Improves performance and accessibility
Adds lazy loading for the YouTube video on the homepage to improve initial page load time.

Replaces direct DOM manipulation with Vue refs for better reactivity and maintainability.

Also, implements a back button component that handles navigating back in history or redirecting to the home page if no history exists.

Adds metadata for improved SEO and accessibility, and implements a CSP to enhance application security.